feat:save and load dashoard from backend
Added new command "save" which will save the dashoard to the backend
This currently only loads/saves to dashboard/1. If the dashboard
is not found, CMPMP will fall back to using localstorage.

When loaded, all changes made are done locally, and "save" needs
to be ran to push the changes to the backend. There are no role
restrictions to who can make local changes, but saving changes
will be determined by the backend server's authorisation